查看离线下载列表/通过 URL 批量添加离线任务/搜索功能/文件管理页「返回」逻辑修改/检测更新...忘了

+++ app/src/main/java/com/jakting/rn6pan/api/data/OfflineListData.kt
@@ -6,13 +6,13 @@ data class OfflineList(
 
 data class OfflineFile(
     var taskIdentity: String,
-    var userIdentity: Int,
-    var createTime: Int,
+    var userIdentity: Long,
+    var createTime: Long,
     var name: String,
     var type: Int,
     var status: Int,
-    var size: Int,
-    var processedSize: Int,
+    var size: Long,
+    var processedSize: Long,
     var progress: Int,
     var errorCode: Int,
     var errorMessage: String,
